Move ReconnectMetrics
